% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lena.R
\name{add_lena_stats}
\alias{add_lena_stats}
\title{Add LENA stats to each interval in a dataframe}
\usage{
add_lena_stats(its_xml, intervals, time_type = c("wav", "wall"))
}
\arguments{
\item{its_xml}{XML object created by \code{rlena::read_its_file}.}

\item{intervals}{A tibble with columns \code{interval_start}, \code{interval_end},
\code{interval_start_wav}, and \code{recording_id} as output by \code{prepare_intervals}.
Can contain other columns.}

\item{time_type}{Either \code{wall} or \code{wav}. If \code{wall}, \code{start} and \code{end} must
contain timestamps with local time. If \code{wav}, \code{start} and \code{end} must contain
the number of milliseconds since the beginning of the wav file. \code{wall} is not
yet implemented.}
}
\value{
Same as \code{intervals} but with three new columns: \code{cvc}, \code{ctc}, and
\code{awc}.
}
\description{
Add LENA stats to each interval in a dataframe
}
\note{
Any its segment overlapping with multiple intervals will count fully
towards all of them.

The difference between this function and \code{calculate_lena_like_stats}
is that the latter calculates exactly the same stats but does it for every
five minute period of wall time (e.g., 13:00-13:05). It should be rewritten
to first create these intervals and then use \code{add_lena_stats} to avoid code
repetition (see issue #).
}
