% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lena.R
\name{add_vtc_stats}
\alias{add_vtc_stats}
\title{Add ctc calculated based on the vtc annotations}
\usage{
add_vtc_stats(intervals, all_rttm)
}
\arguments{
\item{intervals}{A tibble with columns `interval_start`, `interval_end`,
`interval_start_wav`, and `recording_id` as output by `prepare_intervals`.
Can contain other columns.}

\item{all_rttm}{An `all.rttm` file from the VTC output loaded with
`read_rttm`.}
}
\value{
Same as `intervals` but with additional `vtc_ctc` column.
}
\description{
Add ctc calculated based on the vtc annotations
}
\note{
Ported from childrpoject's high-volubility sampler code.
}
