% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bl.R
\name{all_errors}
\alias{all_errors}
\title{A closure to keep track of al the errors from
reading the csv files. There might be a more
elegant way to concatenate and print all
errors from all files, but this seems to do the
job for now.}
\usage{
all_errors()
}
\description{
A closure to keep track of al the errors from
reading the csv files. There might be a more
elegant way to concatenate and print all
errors from all files, but this seems to do the
job for now.
}
