% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdi_wrangling.R
\name{cdi_get_words}
\alias{cdi_get_words}
\title{Select all word item columns for a cdi spreadsheet}
\usage{
cdi_get_words(data, cdi_type = "wg")
}
\arguments{
\item{data}{a dataframe of the original cdi csv}

\item{cdi_type}{Either wg or ws}
}
\value{
New dataframe containing only vocabulary item column
}
\description{
Select all word item columns for a cdi spreadsheet
}
\details{
This function was used to generate the current seedlings cdi spreadsheet
as retrieved by \code{get_cdi_spreadsheet()}.
}
