% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eye_tracking.R
\name{characters_to_factors}
\alias{characters_to_factors}
\title{Convert all character columns to factors}
\usage{
characters_to_factors(df)
}
\arguments{
\item{df}{Dataframe whose columns need conversion.}
}
\value{
Dataframe with all character columns converted to factors.
}
\description{
Convert all character columns to factors
}
