% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{check_for_updates}
\alias{check_for_updates}
\title{Check whether there is a newer version than one requested}
\usage{
check_for_updates(repo, requested_version, tags_already_updated = FALSE)
}
\arguments{
\item{repo}{dataset name: 'all_basiclevel', 'reliability', etc.}

\item{requested_version}{\itemize{
\item the version user requested
}}

\item{tags_already_updated}{boolean, have the repository tags been updated
recently? Avoids unnecessary fetching of the tags when multiple functions
that use tags are called in succession.}
}
\description{
Check whether there is a newer version than one requested
}
