% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_sync.R
\name{checkout_tag}
\alias{checkout_tag}
\title{Download the dataset version specified by a tag}
\usage{
checkout_tag(repo, tag)
}
\arguments{
\item{repo}{dataset name: 'all_basiclevel', 'reliability', etc.}

\item{tag}{tag label as a string}
}
\description{
Download the dataset version specified by a tag
}
\examples{
checkout_tag('all_basiclevel', '0.1.0')
}
