% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bl.R
\name{collect_bl_files}
\alias{collect_bl_files}
\title{Collect basic level files}
\usage{
collect_bl_files(input, type)
}
\arguments{
\item{input}{directory to scan}

\item{type}{datatype of the basic level file ("audio" or "video")}
}
\value{
dataframe with "file" column containing filepaths
}
\description{
Collect basic level files
}
\examples{
x <- collect_bl_files("../some_directory", "video")
x$files
}
