% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bl.R
\name{concat_all_bl}
\alias{concat_all_bl}
\title{Concatenate all the basic level files}
\usage{
concat_all_bl(x, output = NULL)
}
\arguments{
\item{x}{list of by-month basic level tibbles}

\item{output}{if specified, path to the output csv file}
}
\value{
a tibble with all the basic level data
}
\description{
Concatenate all the basic level files
}
\examples{
\dontrun{
x <- concat_month_bl("dir/with/bl/files", "output/folder", "video")
concat_all_bl(x, "all_video.csv")
}
}
