% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bl.R
\name{concat_month_bl}
\alias{concat_month_bl}
\title{Concatenate basic level files by month}
\usage{
concat_month_bl(input, output = NULL, type)
}
\arguments{
\item{input}{directory to scan for basic level files}

\item{output}{directory to write concatenated by-month basic level files to}

\item{type}{basic level datatype ("audio" or "video")}
}
\value{
a list of tibbles, each tibble a month's aggregated basic level
}
\description{
Concatenate basic level files by month
}
