% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{defunct}
\alias{defunct}
\alias{load_tsv}
\alias{fixations_report}
\alias{binifyFixations}
\alias{get_windows}
\alias{FindLowData}
\title{Defunct functions}
\usage{
# Deprecated in 0.22.0 -------------------------------------

load_tsv(fixrep_path, val_guess_max = 1e+05)

fixations_report(
  fixrep_path,
  val_guess_max = 1e+05,
  remove_unfinished = TRUE,
  remove_practice = TRUE
)

binifyFixations(
  gaze,
  binSize = 20,
  keepCols = c("Subject", "TrialNumber", "Target", "T"),
  maxTime = NULL
)

get_windows(
  fix_mes_age,
  bin_size = 20,
  nb_1 = 18,
  short_window_time = DEFAULT_WINDOWS_UPPER_BOUNDS$short,
  med_window_time = DEFAULT_WINDOWS_UPPER_BOUNDS$med,
  long_window_time = DEFAULT_WINDOWS_UPPER_BOUNDS$long
)

FindLowData(gazeData, subsetWin, window_size = NULL, nb_2 = 0, binSize = 20)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#defunct}{\figure{lifecycle-defunct.svg}{options: alt='[Defunct]'}}}{\strong{[Defunct]}}

These functions were removed from blabr. If there's a known replacement,
calling the function will tell you about it.
}
\keyword{internal}
