% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eye_tracking.R
\name{FindLowData}
\alias{FindLowData}
\title{(no docs yet) Mark "low-data" trials with too few bins with fixations}
\usage{
FindLowData(
  gazeData,
  subsetWin,
  window_size = NULL,
  nb_2 = 0,
  binSize = 20,
  propt = "propt",
  timeBin = "timeBin",
  Trial = "Trial",
  SubjectNumber = "SubjectNumber"
)
}
\arguments{
\item{gazeData}{}

\item{subsetWin}{}

\item{window_size}{}

\item{nb_2}{}

\item{binSize}{}

\item{propt}{}

\item{timeBin}{}

\item{Trial}{}

\item{SubjectNumber}{}
}
\description{
(no docs yet) Mark "low-data" trials with too few bins with fixations
}
