% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eye_tracking.R
\name{fixations_report}
\alias{fixations_report}
\title{(no docs yet) Read EyeLink fixation report file}
\usage{
fixations_report(
  fixrep_path,
  val_guess_max = 1e+05,
  remove_unfinished = TRUE,
  remove_practice = TRUE
)
}
\arguments{
\item{val_guess_max}{}

\item{remove_unfinished}{}

\item{remove_practice}{}
}
\description{
(no docs yet) Read EyeLink fixation report file
}
