% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eye_tracking.R
\name{fixations_to_timeseries}
\alias{fixations_to_timeseries}
\title{(no docs yet) Convert a fixations table to an evenly spaced timeseries}
\usage{
fixations_to_timeseries(
  gaze,
  binSize = 20,
  keepCols = c("Subject", "TrialNumber", "Target", "T"),
  maxTime = NULL
)
}
\arguments{
\item{maxTime}{}
}
\description{
(no docs yet) Convert a fixations table to an evenly spaced timeseries
}
