% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blab_participants.R
\name{get_blab_excluded_participants}
\alias{get_blab_excluded_participants}
\title{Get the list of excluded participants from the blab_wide tracking sheet}
\usage{
get_blab_excluded_participants()
}
\value{
A dataframe of exluded participants, their local subject ids and notes
}
\description{
Get the list of excluded participants from the blab_wide tracking sheet
}
