% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blab_participants.R
\name{get_blab_participants}
\alias{get_blab_participants}
\title{Get clean blab-wide participants data from participants tracking sheet and standardize the local ids}
\usage{
get_blab_participants()
}
\value{
a dataframe of all blab participants with standardized local ids
}
\description{
Get clean blab-wide participants data from participants tracking sheet and standardize the local ids
}
