% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{get_blab_share_path}
\alias{get_blab_share_path}
\title{Find BLab share}
\usage{
get_blab_share_path(check_exists = TRUE)
}
\arguments{
\item{check_exists}{Should the function check that the path exists
(accessible)}
}
\value{
Path to the BLab share
}
\description{
Find BLab share
}
\examples{
\dontrun{
blab_share <- get_blab_share_path()
}
}
