% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_sync.R
\name{get_current_commit_date}
\alias{get_current_commit_date}
\title{Get the date of the currently checked out commit}
\usage{
get_current_commit_date(repo)
}
\arguments{
\item{repo}{dataset name: 'all_basiclevel', 'reliability', etc.}
}
\value{
A string with the date.
}
\description{
Get the date of the currently checked out commit
}
\examples{

\dontrun{
get_current_commit_date('all_basiclevel')
}
}
