% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_sync.R
\name{get_current_tag}
\alias{get_current_tag}
\title{Find tag label of the currently checked out commit.}
\usage{
get_current_tag(repo, tags_already_updated = FALSE)
}
\arguments{
\item{repo}{dataset name: 'all_basiclevel', 'reliability', etc.}

\item{tags_already_updated}{boolean, have the repository tags been updated
recently? Avoids unnecessary fetching of the tags when multiple functions
that use tags are called in succession.}
}
\value{
The latest version tag as a string
}
\description{
Find tag label of the currently checked out commit.
}
\examples{
\dontrun{
all_bl <- get_all_basiclevel(version = '0.0.7')
current_tag <- get_current_tag('all_basiclevel')
stopifnot(current_tag == '0.0.7')
}
}
