% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eye_tracking.R
\name{get_late_target_onset}
\alias{get_late_target_onset}
\title{(no docs yet) Find late target onsets and create a doc to check and potentially correct them}
\usage{
get_late_target_onset(
  data,
  max_time = 6000,
  study = "eye_tracking_study",
  output_dir = "../data/",
  out_csv = FALSE
)
}
\description{
(no docs yet) Find late target onsets and create a doc to check and potentially correct them
}
