% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_sync.R
\name{get_latest_tag}
\alias{get_latest_tag}
\title{Finds the latest version tag}
\usage{
get_latest_tag(repo, tags_already_updated = FALSE)
}
\arguments{
\item{repo}{dataset name: 'all_basiclevel', 'reliability', etc.}

\item{tags_already_updated}{boolean, have the repository tags been updated
recently? Avoids unnecessary fetching of the tags when multiple functions
that use tags are called in succession.}
}
\value{
The latest version tag as a string
}
\description{
Finds the latest version tag
}
\examples{
\dontrun{
get_latest_tag('all_basiclevel')
}
}
