% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_latest_version}
\alias{get_latest_version}
\title{Find latest version available for downloading?}
\usage{
get_latest_version(repo, tags_already_updated = FALSE)
}
\arguments{
\item{repo}{dataset name: 'all_basiclevel', 'reliability', etc.}

\item{tags_already_updated}{boolean, have the repository tags been updated
recently? Avoids unnecessary fetching of the tags when multiple functions
that use tags are called in succession.}
}
\value{
the version string
}
\description{
Find latest version available for downloading?
}
\examples{
get_latest_version('all_basiclevel')
}
