% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{get_pn_opus_path}
\alias{get_pn_opus_path}
\title{Find PN-OPUS}
\usage{
get_pn_opus_path(check_exists = TRUE)
}
\value{
Path to the PN-OPUS shared drive
}
\description{
Find PN-OPUS
}
\examples{
pn_opus_path <- get_pn_opus_path()
}
