% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdi_wrangling.R
\name{get_r01_cdi}
\alias{get_r01_cdi}
\title{Load CDI output from all RO1 studies and SemPhonD}
\usage{
get_r01_cdi(
  table = c("summary", "wordlevel", "raw"),
  study = NULL,
  withDemographic = FALSE,
  justWord = TRUE,
  version = NULL
)
}
\arguments{
\item{table}{Which subset of the output to include?
All table include id-related columns: "study_name", "subject_id", "repeat_num".
\itemize{
\item "summary" (default): each row is one cdi form, including aggregated score
for each form (e.g. total words produced, total number of first gestures, etc.).
\item "wordlevel": each row is an item on one cdi form, with an "item" column
for the name of the item and a "response" column.
\item "raw": this is the rawest form of the data as downloaded from webcdi, with
id-related columns added, each row is one cdi form, including columns for
each item and columns for aggregated score,
as well as metadata columns (e.g. date created, webcdi link, admin id, etc.).
Is not affected by \code{withDemographic} and \code{justWord} parameters.
}}

\item{study}{Which study(s) to include among (\code{WFR}, \code{CLF}, \code{PBS}, \code{MISCOM},
\code{ProsPr}, \code{PreFunc}, \code{SemPhonD}, or \code{SPARKED})? Provide either a string (if only
selecting one study) or a vector of strings (for multiple studies)
If \code{NULL} (default), will include all studies.}

\item{withDemographic}{Should the data include answers to demographic questions
(e.g. sibling count, birth order, caregive info, etc.)? These columns are not
treated as item, so if you select \code{wordlevel_long} table, the demographic
answers will be included in every row. If you select \code{raw} table, demographic
columns will always be included. \code{FALSE} is default.}

\item{justWord}{Should the data include only vocabulary checklist item? If
\code{TRUE} (default), returns only \verb{Words Produced} and \verb{Words Understood} related
columns for \code{summary} table and vocabulary items for \code{wordlevel} table. If
\code{FALSE}, will include gestures-related items for \code{WG} form and sentence-related
items for \code{WS} form. If you select \code{raw} table, all items will be included
regardless of this variable.}

\item{version}{version tag to checkout}
}
\description{
Load CDI output from all RO1 studies and SemPhonD
}
\details{
Clone BLAB-private \href{https://github.com/bergelsonlab/r01_cdi_spreadsheet.git}{r01_cdi_spreadsheet}
repo to \verb{~/BLAB_DATA} once before using this function.
}
