% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_reliability}
\alias{get_reliability}
\title{Get the reliability spreadsheets}
\usage{
get_reliability(av, month, version = NULL)
}
\arguments{
\item{av}{either "audio" or "video"}

\item{month}{the month as a string, e.g. "06"}

\item{version}{version tag to checkout}
}
\value{
dataframe with data for the requested modality and month
}
\description{
Get the reliability spreadsheets
}
\examples{
audio_06_rel <- get_reliability("audio", "06")
}
