% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_sync.R
\name{get_repo_path}
\alias{get_repo_path}
\title{Absolute path to a dataset folder}
\usage{
get_repo_path(repo)
}
\arguments{
\item{repo}{dataset name: 'all_basiclevel', 'reliability', etc.}
}
\value{
absolute path as a string
}
\description{
Absolute path to a dataset folder
}
\examples{
get_repo_path('all_basiclevel')
}
