% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdi_wrangling.R
\name{get_seedlings_cdi}
\alias{get_seedlings_cdi}
\title{Get the seedlings CDI table from the BLAB_DATA \href{https://github.com/BergelsonLab/cdi_spreadsheet}{repo}
using \code{get_cdi_spreadsheet()}, but wrangled to include summary values and
with options for which form to retrieve the dataframe.
The norm conversion table was downloaded from https://github.com/langcog/wordbank-shiny/tree/main/apps/scoring/norms/percentiles/English\%20Percentiles}
\usage{
get_seedlings_cdi(
  version = NULL,
  table = c("summary", "wordlevel", "raw"),
  justWord = TRUE
)
}
\arguments{
\item{version}{version tag to checkout}

\item{table}{Which subset of the output to include?
All table include demographic related items: "ResponseID", "SubjectNumber",
"subj", "month", "Date_Completed", "AgeMonthUncorrected",
"SeedlingsFinalSample", "Relation_to_child", "Child_gender"
\itemize{
\item "summary" (default): each row is one CDI form, including aggregated score
for for each section of the CDI (CDIcomp, game gestures, later gestures). If
a section does not have an aggregate (i.e. first signs, starting to talk),
will include them as separate items with binary values (1 or 0).
\item "wordlevel": each row is an item on one cdi form, with an "item" column
for the name of the item and a "response" column.
\item "raw": this is the rawest form of the data as saved in BLAB_DATA, with one
row for each CDI form, including demographic columns, items columns, and
CDIcomp, CDIprod. Is not affected by \code{justWord} parameters.
}}

\item{justWord}{Should the data include only vocabulary checklist item? If
\code{TRUE} (default), returns only \code{CDIcomp} and \code{CDIprod} related  for \code{summary}
table and only vocabulary items for \code{wordlevel} table. If
\code{FALSE}, will include gestures-related items.  If you select \code{raw} table, all
items will be included regardless of this variable.}
}
\value{
A dataframe of the wrangled CDI output according to the parameters
}
\description{
Get the seedlings CDI table from the BLAB_DATA \href{https://github.com/BergelsonLab/cdi_spreadsheet}{repo}
using \code{get_cdi_spreadsheet()}, but wrangled to include summary values and
with options for which form to retrieve the dataframe.
The norm conversion table was downloaded from https://github.com/langcog/wordbank-shiny/tree/main/apps/scoring/norms/percentiles/English\%20Percentiles
}
