% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_versions}
\alias{get_versions}
\title{List available versions}
\usage{
get_versions(repo, tags_already_updated = FALSE)
}
\arguments{
\item{repo}{dataset name: 'all_basiclevel', 'reliability', etc.}

\item{tags_already_updated}{boolean, have the repository tags been updated
recently? Avoids unnecessary fetching of the tags when multiple functions
that use tags are called in succession.}
}
\value{
the version string
}
\description{
List available versions
}
\examples{
get_versions('seedlings-nouns')
}
