% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdi_wrangling.R
\name{get_vocab_score}
\alias{get_vocab_score}
\title{Calculate the vocabulary checklist score of a cdi spreadsheet}
\usage{
get_vocab_score(data, cdi_type, remove_incomplete = T)
}
\arguments{
\item{data}{a dataframe of the original cdi csv}

\item{cdi_type}{Either wg or ws}

\item{remove_incomplete}{whether to remove any incomplete cdi forms}
}
\value{
New dataframe with vocab score
}
\description{
Calculate the vocabulary checklist score of a cdi spreadsheet
}
\details{
This function was used to generate the current seedlings cdi spreadsheet
as retrieved by \code{get_cdi_spreadsheet()}.
}
