% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lena.R
\name{get_vtc_speaker_stats}
\alias{get_vtc_speaker_stats}
\title{Calculate per-speaker statistics based on the VTC output}
\usage{
get_vtc_speaker_stats(intervals, all_rttm)
}
\arguments{
\item{intervals}{A tibble with columns `interval_start`, `interval_end`,
`interval_start_wav`, and `recording_id` as output by `prepare_intervals`.
Can contain other columns.}

\item{all_rttm}{An `all.rttm` file from the VTC output loaded with
`read_rttm`.}
}
\value{
A tibble with `interval_start` and `interval_end` columns for each
interval in `intervals` and three new columns: `voice_type` and `duration`
and `count` of VTC annotations. The `duration` is in seconds.
}
\description{
Calculate per-speaker statistics based on the VTC output
}
