% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{handle_dataset_version}
\alias{handle_dataset_version}
\title{Handles the version.}
\usage{
handle_dataset_version(
  repo,
  version = NULL,
  tags_already_updated = FALSE,
  check_for_updates = TRUE
)
}
\arguments{
\item{version}{version tag to checkout}
}
\value{
version string
}
\description{
If the version isn't specified, finds the newest version and warns that not
specifying the version might not be a good idea.
If it is specified, notifies if there is a newer version available.
}
