% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bl.R
\name{join_full_audio_video}
\alias{join_full_audio_video}
\title{Join the audio and video dataframes into one big dataframe}
\usage{
join_full_audio_video(
  audiostats,
  videostats,
  output_name = NULL,
  keep_na = FALSE,
  keep_comments = FALSE
)
}
\arguments{
\item{audiostats}{the full audio basic level dataframe}

\item{videostats}{the full video basic level dataframe}

\item{output}{if specified, the output name for csv and feather files (without extension)}
}
\value{
the joined dataframe
}
\description{
Join the audio and video dataframes into one big dataframe
}
\examples{
\dontrun{
audiomonths <- concat_month_bl("../basiclevel_dataprep/all_bl/", "output/", "audio")
audiostats <- concat_all_bl(audiomonths, "all_audio.csv")
videomonths <- concat_month_bl("../basiclevel_dataprep/all_bl/", "output/", "video")
videostats <- concat_all_bl(videomonths, "all_video.csv")
joined_data <- join_full_audio_video(audiostats, videostats)
}
}
