% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eye_tracking.R
\name{keypress_issues}
\alias{keypress_issues}
\title{(no docs yet) Find key press issues and create a doc to check and potentially correct them}
\usage{
keypress_issues(
  data,
  study = "eye_tracking_study",
  practice_trials = c("p1", "p2", "p3", "p4"),
  output_dir = "../data/",
  out_csv = FALSE
)
}
\description{
(no docs yet) Find key press issues and create a doc to check and potentially correct them
}
