% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eye_tracking.R
\name{keypress_retrieved}
\alias{keypress_retrieved}
\title{(no docs yet) Load file with manually corrected key presses or late target onsets}
\usage{
keypress_retrieved(
  filename,
  drop_list = c("video_pop_time", "video_targetonset", "notes")
)
}
\description{
(no docs yet) Load file with manually corrected key presses or late target onsets
}
