% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_rename}
\alias{list_rename}
\title{\code{dplyr::rename} for lists}
\usage{
list_rename(.x, ..., .strict = TRUE)
}
\arguments{
\item{.x}{A list.}

\item{...}{Any number of \code{new_name = old_name} pairs.}
}
\description{
\code{dplyr::rename} for lists
}
