% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_basic_level.r
\name{make_new_global_basic_level}
\alias{make_new_global_basic_level}
\title{Create all_basiclevel_na with added global_bl column}
\usage{
make_new_global_basic_level(
  all_basiclevel_version = NULL,
  global_bl_mappings_version = NULL
)
}
\arguments{
\item{all_basiclevel_version}{Version tag of the \code{all_basiclevel} repository.
Use for testing only.}

\item{global_bl_mappings_version}{Version tag of the repository with the
mappings used to match tokens with their global basic levels. Use for
testing only.

Loads the latest versions of:
\itemize{
\item all_basiclevel_na from the "all_basiclevel" repository,
\item object_dict and annotid_disambiguation from the "global_basic_level"
repository
}

And then tries to add a global_bl column to all_basiclevel_na if it can.
If it can, it will return all_basiclevel_na with an added a column.

If it can't, it will throw an error and provide instructions how to update
the global basic level mappings.}
}
\description{
Create all_basiclevel_na with added global_bl column
}
