% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eye_tracking.R
\name{merge_split_reports}
\alias{merge_split_reports}
\title{Merge tables from split fixation and message reports}
\usage{
merge_split_reports(fix_rep_data, mes_rep_data)
}
\arguments{
\item{fix_rep_data}{List of tables as returned by \code{split_fixation_report}.}

\item{mes_rep_data}{List of tables as returned by \code{split_message_report}.}
}
\description{
The main point of merging is to have one list and no repeating data. For
example, we don't have two "trials" tables in memory where the one coming
from the message report has more trials - we want them merged.
}
