% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lena.R
\name{prepare_intervals}
\alias{prepare_intervals}
\title{Prepare intervals to potentially be annotated later}
\usage{
prepare_intervals(its_xml, duration)
}
\arguments{
\item{its_xml}{XML object created by `rlena::read_its_file`.}

\item{duration}{Interval duration supported by `lubridate::period`, e.g.,
'2 mins'.}
}
\value{
A tibble with
}
\description{
Various metrics can then be calculated for each intervals and the "best"
can then be annotated.
}
