% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lena.R
\name{prepare_intervals_for_sampling}
\alias{prepare_intervals_for_sampling}
\title{Prepare intervals for sampling}
\usage{
prepare_intervals_for_sampling(intervals, size, period, allow_fewer)
}
\arguments{
\item{intervals}{a tibble with interval_start and interval_end datetime
columns}

\item{size}{required sample size}

\item{period}{What period (e.g., '5 mins') is the main interval size? Only
intervals of this duration will be sampled.}

\item{allow_fewer}{Is it OK if there are fewer than size intervals?}
}
\value{
intervals with incomplete intervals removed
}
\description{
- remove incomplete intervals,
- check that there are still some left after that,
- check that there are at least `size` intervals, unless `allow_fewer` is
  `TRUE`
}
\keyword{internal}
