% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bl.R
\name{process_concat_bl}
\alias{process_concat_bl}
\title{Post-processing for the full basic level dataframe}
\usage{
process_concat_bl(x)
}
\arguments{
\item{x}{full (audio or video) basic level dataframe}
}
\value{
processed basic level dataframe
}
\description{
Post-processing includes adding a subject and month column, whether
it's an audio or video, and reformatting timestamps.
}
\details{
The function will figure out whether it's an audio or video dataframe.
It must be one or the other though. It will not handle a joined audio+video
frame.
}
