% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/babar.R
\name{read_babar}
\alias{read_babar}
\title{Read in a babar output csv file, or a folder of babar output csv files, and
compile into one dataframe}
\usage{
read_babar(filepath, batch)
}
\arguments{
\item{filepath}{path to either a babar csv file or a folder containing multiple
babar csv files}

\item{batch}{TRUE if reading a folder, FALSE if reading only one file}
}
\value{
A combined dataframe of all babar csv file read, with new columns for
duration and recording_id without .eaf extension
}
\description{
Read in a babar output csv file, or a folder of babar output csv files, and
compile into one dataframe
}
