% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eye_tracking.R
\name{read_report}
\alias{read_report}
\title{Read EyeLink fixation/message report file}
\usage{
read_report(
  report_path,
  guess_max = 1e+05,
  remove_unfinished = TRUE,
  remove_practice = TRUE
)
}
\arguments{
\item{report_path}{Report file path.}

\item{guess_max}{Passed to \code{readr::read_tsv}. Default is 100000. Decreasing
this number can speed up reading, just make sure there were no problems
during reading by looking out for warning from \code{readr} or by passing the
output dataframe to \code{readr::problems}.}

\item{remove_unfinished}{Removes lines where the \code{order} column is NA. Does
not work in the current version because no value are interpreted as NA.}

\item{remove_practice}{Only keeps rows where the \code{practice} column is "n".}
}
\description{
Read EyeLink fixation/message report file
}
\keyword{internal}
