% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rttm.R
\name{read_rttm}
\alias{read_rttm}
\title{Read an .rttm file produced by Voice Type Classifier}
\usage{
read_rttm(rttm_path)
}
\arguments{
\item{rttm_path}{Path to the .rttm file to be read.}
}
\value{
a tibble with onset and duration as numeric and other columns as
character
}
\description{
Read an .rttm file produced by Voice Type Classifier
}
