% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seedlings.R
\name{read_seedlings_audio_annotations}
\alias{read_seedlings_audio_annotations}
\title{Read audio sparse code csv from the Seedlings annotations.}
\usage{
read_seedlings_audio_annotations(file_path, parse_timestamp_audio = T)
}
\arguments{
\item{file_path}{Path to the sparse code csv file.}

\item{parse_timestamp_audio}{Should the timestamp column be split into the
"onset" and "offset" integer columns?}
}
\value{
A tibble with all the columns in the file at \code{file_path} read as
character vector. If \code{parse_timestamp_audio} is true, there will be two
additional integer columns: \code{onset} and \code{offset}.
}
\description{
Read audio sparse code csv from the Seedlings annotations.
}
