% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lena.R
\name{sample_intervals_periodically}
\alias{sample_intervals_periodically}
\title{Sample intervals periodically, e.g. every hour}
\usage{
sample_intervals_periodically(intervals, duration, period)
}
\arguments{
\item{intervals}{A tibble with columns \code{interval_start}, \code{interval_end},
\code{interval_start_wav}, and \code{recording_id} as output by \code{prepare_intervals}.
Can contain other columns.}

\item{duration}{Interval duration supported by \code{lubridate::period}, e.g.,
'2 mins'.}

\item{period}{Period at the end of which we should sample intervals,
e.g., '1 hour'}
}
\value{
A subsample of intervals.
}
\description{
Sample intervals periodically, e.g. every hour
}
