% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lena.R
\name{sample_intervals_with_highest}
\alias{sample_intervals_with_highest}
\title{Sample intervals that are highest on a given metric}
\usage{
sample_intervals_with_highest(intervals, column, size, period,
  allow_fewer = FALSE)
}
\arguments{
\item{intervals}{a tibble with interval_start and interval_end datetime
columns}

\item{column}{Name of the column whose value is to be maximized.}

\item{size}{required sample size}

\item{period}{What period (e.g., '5 mins') is the main interval size? Only
intervals of this duration will be sampled.}

\item{allow_fewer}{Is it OK if there are fewer than size intervals?}
}
\description{
Sample intervals that are highest on a given metric
}
