% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eye_tracking.R
\name{split_fixation_report}
\alias{split_fixation_report}
\title{Split a fixation report into a list of hierarchical dataframes}
\usage{
split_fixation_report(
  report,
  drop_empty_columns = TRUE,
  drop_saccades = TRUE,
  drop_prev_and_next = TRUE
)
}
\arguments{
\item{report}{Fixation report dataframe as created by \code{read_fixation_report}.}

\item{drop_saccades, }{drop_prev_and_next Logical. If TRUE, columns related to
saccades and previous/next fixations/saccades will be dropped.}
}
\description{
Split a fixation report into a list of hierarchical dataframes
}
