% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eye_tracking.R
\name{split_message_report}
\alias{split_message_report}
\title{Split a message report into a list of hierarchical dataframes}
\usage{
split_message_report(report, drop_empty_columns = TRUE)
}
\arguments{
\item{report}{Message report dataframe as created by \code{read_message_report}.}

\item{drop_empty_columns}{Logical. If TRUE, columns that are all NA will be
dropped.}
}
\description{
Split a message report into a list of hierarchical dataframes
}
