% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eye_tracking.R
\name{split_report}
\alias{split_report}
\title{Split a fixation or message report into hierarchical dataframes}
\usage{
split_report(report, drop_empty_columns = TRUE)
}
\arguments{
\item{drop_empty_columns}{Logical. If TRUE, columns that are all NA will be
dropped.}
}
\description{
Split a fixation or message report into hierarchical dataframes
}
\keyword{internal}
