% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eye_tracking.R
\name{tag_low_data_trials}
\alias{tag_low_data_trials}
\title{Mark "low-data" trials}
\usage{
tag_low_data_trials(
  fixation_timeseries,
  window_column,
  t_start,
  t_end = NULL,
  t_step = 20,
  min_fraction = 1/3
)
}
\arguments{
\item{fixation_timeseries}{A fixations dataframe that is required to minimally contain these columns:
\itemize{
\item \code{recording_id} and \code{trial_index} to uniquely trials.
\item \code{is_good_timepoint} - only rows with TRUE in this column will be counted as having data.
\item Column specified by \code{window_column}.
}}

\item{window_column}{Name of the column that indicates (using factor label "Y") bins that belong to the window which will be tested for insufficient data.}

\item{t_start}{Lower bound of the window of interest in milliseconds from the target onset.}

\item{t_end}{Upper bound of the window of interest in milliseconds from the target onset. If a non-default upper bound was used during the assignment of bins to windows in the \code{get_windows} call, then that value MUST be supplied here. If NULL (default), the default corresponding to the \code{subsetWin} will be used.}

\item{t_step}{The time step in ms. MUST match the one used in \code{fixations_to_timeseries}.}

\item{min_fraction}{Minimum fraction of the window that must contain fixations on either area of interest for the trial to be considered "high-data". Default is 1/3.}
}
\value{
The input dataframe with boolean \code{is_trial_low_data} column added.
}
\description{
Trials are considered to be "low-data" if less than \code{min_fraction} (1/3 by default) of the window of interest contains fixations on either area of interest.
Time points that  count are the ones for which:
\itemize{
\item \code{is_good_timepoint} is TRUE. Typically, created with something like \verb{mutate(is_good_timepoint = aoi \%in\% c('TARGET', 'DISTRACTROR')}.
The definition can differ between studies.
\item Column specified in \code{window_column} is 'Y'.
}
}
