% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_sync.R
\name{update_tags}
\alias{update_tags}
\title{Fetches dataset tags from GitHub}
\usage{
update_tags(repo)
}
\arguments{
\item{repo}{dataset name: 'all_basiclevel', 'reliability', etc.}
}
\description{
Fetches dataset tags from GitHub
}
\examples{
\dontrun{
update_tags('all_basiclevel')
}
}
