% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blab_participants.R
\name{wrangle_blab_participants}
\alias{wrangle_blab_participants}
\title{Standardize and clean up local subject IDs for blab participants}
\usage{
wrangle_blab_participants(df)
}
\arguments{
\item{df}{A data frame containing participant information, loaded in from the blab-wide participant spreadsheet}
}
\value{
a new data frame with standardized IDs
}
\description{
Standardize and clean up local subject IDs for blab participants
}
