% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdi_wrangling.R
\name{wrangle_web_cdi}
\alias{wrangle_web_cdi}
\title{Clean up raw CDI output (item + summary) for any project administered through
WebCDI.}
\usage{
wrangle_web_cdi(
  cdi_df,
  form = c("WG", "WS"),
  table = c("summary", "wordlevel", "raw"),
  withDemographic = FALSE,
  justWord = TRUE,
  rename = TRUE
)
}
\arguments{
\item{form}{Which kind of cdi form is this (\code{WG} or \code{WS}).
Will only include columns from each form.}

\item{table}{Which subset of the output to include?
All table include id-related columns: "study_name", "subject_id", "repeat_num".
\itemize{
\item "summary" (default): each row is one cdi form, including aggregated score
for each form (e.g. total words produced, total number of first gestures, etc.).
\item "wordlevel": each row is an item on one cdi form, with an "item" column
for the name of the item and a "response" column.
\item "raw": this is the rawest form of the data as downloaded from webcdi, with
id-related columns added, each row is one cdi form, including columns for
each item and columns for aggregated score,
as well as metadata columns (e.g. date created, webcdi link, admin id, etc.).
Is not affected by \code{withDemographic} and \code{justWord} parameters.
}}

\item{withDemographic}{Should the data include answers to demographic questions
(e.g. sibling count, birth order, caregive info, etc.)? These columns are not
treated as item, so if you select \code{wordlevel_long} table, the demographic
answers will be included in every row. If you select \code{raw} table, demographic
columns will always be included. \code{FALSE} is default.}

\item{justWord}{Should the data include only vocabulary checklist item? If
\code{TRUE} (default), returns only \verb{Words Produced} and \verb{Words Understood} related
columns for \code{summary} table and vocabulary items for \code{wordlevel} table. If
\code{FALSE}, will include gestures-related items for \code{WG} form and sentence-related
items for \code{WS} form. If you select \code{raw} table, all items will be included
regardless of this variable.}

\item{rename}{For \code{WS} form, do we standardize the name of the
word-related summary item to be the same as \code{WG} form. If \code{TRUE} (default), the
three summary columns in WS named \verb{Total Produced}, \verb{Total Produced Percentile-sex},
and \verb{Total Produced Percentile-both} will  changed to \verb{Words Produced},
\verb{Words Produced Percentile-sex} and \verb{Words Produced Percentile-both}. Otherwise,
they will remain unchanged}

\item{filepath}{Dataframe of the csv file downloaded from WebCDI
(you should use the csv with both item level and sumary data)}
}
\value{
A dataframe of the wrangled CDI output according to the parameters
}
\description{
Clean up raw CDI output (item + summary) for any project administered through
WebCDI.
}
