% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rttm.R
\name{write_rttm}
\alias{write_rttm}
\title{Write a tibble to an .rttm file}
\usage{
write_rttm(rttm_df, rttm_path)
}
\arguments{
\item{rttm_df}{A dataframe with rttm data as produced by \code{read_rttm}.}

\item{rttm_path}{A path to write to.}
}
\value{
Returns input \code{rttm_df}
}
\description{
Write a tibble to an .rttm file
}
