% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seedlings.R
\name{write_seedlings_audio_annotations}
\alias{write_seedlings_audio_annotations}
\title{Write the seedlings audio annotation tibble back to the csv file}
\usage{
write_seedlings_audio_annotations(annotations, file_path)
}
\arguments{
\item{annotations}{A tibble such as read by read_seedlings_audio_annotations.}

\item{file_path}{File path to write to.}
}
\value{
Input \code{annotations} with any extra columns removed and the "pho"
column having NA filled with empty strings - the tibble that is fed to
\code{dplyr::write_csv}.
}
\description{
Write the seedlings audio annotation tibble back to the csv file
}
