% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{write_versions_info}
\alias{write_versions_info}
\title{Write version information}
\usage{
write_versions_info(to_file = "versions/session-info.txt")
}
\arguments{
\item{to_file}{}
}
\description{
This function will write version information about all the installed R
packaged and about all the relevant external software.
}
\details{
Run this function in the root of your project before any collaborator
(whether internal or external) will work on the project.
}
