% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lena.R
\name{add_lena_stats}
\alias{add_lena_stats}
\title{Add LENA stats to each interval in a dataframe}
\usage{
add_lena_stats(its_xml, intervals, time_type)
}
\arguments{
\item{its_xml}{XML object created by `rlena::read_its_file`.}

\item{intervals}{Tibble that contain columns `start` and `end`}

\item{time_type}{Either `wall` or `wav`. If `wall`, `start` and `end` must
contain timestamps with local time. If `wav`, `start` and `end` must contain
the number of milliseconds since the beginning of the wav file. `wall` is not
yet implemented.}
}
\value{
Same as `intervals` but with three new columns: `cvc`, `ctc`, and
`awc`.
}
\description{
Add LENA stats to each interval in a dataframe
}
\note{
Any its segment overlapping with multiple intervals will count fully
towards all of them.

The difference between this function and `calculate_lena_like_stats`
is that the latter calculates exactly the same stats but does it for every
five minute period of wall time (e.g., 13:00-13:05). It should be rewritten
to first create these intervals and then use `add_lena_stats` to avoid code
repetition (see issue #).
}
