% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lena.R
\name{add_vtc_stats}
\alias{add_vtc_stats}
\title{Add ctc calculated based on the vtc annotations}
\usage{
add_vtc_stats(intervals, all_rttm)
}
\arguments{
\item{intervals}{A tibble with columns depending on \code{time_type}:
\itemize{
\item For \code{wall}: \code{interval_start}, \code{interval_end} (POSIXct), \code{interval_start_wav} (ms)
\item For \code{wav}: \code{interval_start_wav}, \code{interval_end_wav} (both in ms since wav start)
Can contain other columns.
}}

\item{all_rttm}{An \code{all.rttm} file from the VTC output loaded with
\code{read_rttm}.}
}
\value{
Same as \code{intervals} but with additional \code{vtc_ctc} column.
}
\description{
Add ctc calculated based on the vtc annotations
}
\note{
Ported from childrpoject's high-volubility sampler code.
}
