% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eye_tracking.R
\name{binifyFixations}
\alias{binifyFixations}
\title{(no docs yet) Bin fixations}
\usage{
binifyFixations(
  gaze,
  binSize = 20,
  keepCols = c("Subject", "TrialNumber", "Target", "T"),
  maxTime = NULL
)
}
\arguments{
\item{gaze}{}

\item{binSize}{}

\item{keepCols}{}

\item{maxTime}{}
}
\description{
(no docs yet) Bin fixations
}
