% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lena.R
\name{calculate_lena_like_stats}
\alias{calculate_lena_like_stats}
\title{Calculate stats similar to those in LENA's 5min.csv files}
\usage{
calculate_lena_like_stats(its_xml, duration)
}
\arguments{
\item{its_xml}{XML object created by `rlena::read_its_file`.}

\item{duration}{Interval duration supported by `lubridate::period`, e.g.,
'2 mins'.}
}
\value{
a tibble with at least these four columns: interval_start,
interval_end, AWC.Actual, CTC.Actual, CWC.Actual
}
\description{
Calculate stats similar to those in LENA's 5min.csv files
}
