% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_all_basiclevel}
\alias{get_all_basiclevel}
\title{Get the all_basiclevel data}
\usage{
get_all_basiclevel(version = NULL, type = "feather")
}
\arguments{
\item{version}{version tag to checkout}

\item{type}{"feather" or "csv". defaults to "feather"}
}
\value{
a dataframe containing the all_basicalevel data
}
\description{
Get the all_basiclevel data
}
\examples{

# get version with a specific version tag
all_bl <- get_all_basiclevel(version='0.1.0')
}
