% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_all_basiclevel}
\alias{get_all_basiclevel}
\title{Get the all_basiclevel data}
\usage{
get_all_basiclevel(branch = NULL, commit = NULL, type = "feather")
}
\arguments{
\item{branch}{git branch to pull from}

\item{commit}{git commit to pull}

\item{type}{"feather" or "csv". defaults to "feather"}
}
\value{
a tibble containing the all_basicalevel data
}
\description{
Get the all_basiclevel data
}
\examples{
all_bl <- get_all_basiclevel()
}
