% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_all_basiclevel}
\alias{get_all_basiclevel}
\title{Get the all_basiclevel data from the all_basiclevel repo}
\usage{
get_all_basiclevel(version = NULL, drop_basic_level_na = TRUE)
}
\arguments{
\item{version}{version tag to checkout}

\item{drop_basic_level_na}{whether to use the "*_NA" version that has all the
nouns, including those whose "basic_level" is set to "NA"
(drop_basic_level_na = FALSE) or the standard version that does not include
them.}
}
\value{
a dataframe containing the all_basiclevel data
}
\description{
Get the all_basiclevel data from the all_basiclevel repo
}
\examples{

# get version with a specific version tag
all_bl <- get_all_basiclevel(version='0.3.2')
}
