% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eye_tracking.R
\name{get_clf_eyetracking}
\alias{get_clf_eyetracking}
\title{The wrangled eyetracking spreadsheet for CLF eyetracking components of all studies (CLF, PBS, Prefunc, ProsPr, WFR, biWFR).}
\usage{
get_clf_eyetracking(version = NULL)
}
\arguments{
\item{version}{version tag to checkout}
}
\value{
a dataframe containing the fixation as time series, excluding invalid trials
}
\description{
The wrangled eyetracking spreadsheet for CLF eyetracking components of all studies (CLF, PBS, Prefunc, ProsPr, WFR, biWFR).
}
\examples{
motor <- get_motor_spreadsheet(version = '0.0.2')
}
