% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/babar.R
\name{get_consonant_inventory}
\alias{get_consonant_inventory}
\title{Collecting the consonant inventory of each recording. Glides are not included.}
\usage{
get_consonant_inventory(df, minimum_count = 50)
}
\arguments{
\item{df}{babar dataframe}

\item{minimum_count}{the minimum number of occurences of a phoneme per recording
to be counted as a phoneme. Default is 50.}
}
\value{
A dataframe with these new metrics as new columns
}
\description{
Collecting the consonant inventory of each recording. Glides are not included.
}
