% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_dataset_version}
\alias{get_dataset_version}
\title{Get the commit hash and date of a dataset}
\usage{
get_dataset_version(dataset)
}
\arguments{
\item{dataset}{- name of the dataset}
}
\description{
For a given dataset, returns hash and date of the currently checked out
commit. This can be different from the commit used in the parent BLAB_DATA
repo.
}
\examples{
all_bl_version <- get_dataset_version('all_basiclevel')
print(all_bl_version$commit)
print(all_bl_version$date)
}
