% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_df_file}
\alias{get_df_file}
\title{Downloads and then loads a csv/feather file from a specified version of a
dataset}
\usage{
get_df_file(repo, filename, version = NULL, col_types = NULL)
}
\arguments{
\item{repo}{dataset name: 'all_basiclevel', 'reliability', etc.}

\item{filename}{name of a csv/feather file}

\item{version}{version tag to checkout}

\item{col_types}{Passed to `readr::read_csv` when filename ends with ".csv".}
}
\value{
tibble for feather files, data.frame for csv files
}
\description{
Downloads and then loads a csv/feather file from a specified version of a
dataset
}
\examples{

\dontrun{
get_df_file('all_basiclevel', 'all_basiclevel.csv', version = '0.1.0')
}
}
