% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/babar.R
\name{get_inventory}
\alias{get_inventory}
\title{Collecting the phonetic inventory of each recording.
The inventory is a space separated string.}
\usage{
get_inventory(df, minimum_count = 50)
}
\arguments{
\item{df}{babar dataframe}

\item{minimum_count}{the minimum number of occurences of a phoneme per recording
to be counted as a phoneme. Default is 50.}
}
\value{
A dataframe with these new metrics as new columns
}
\description{
Collecting the phonetic inventory of each recording.
The inventory is a space separated string.
}
