% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eye_tracking.R
\name{get_mesrep}
\alias{get_mesrep}
\title{(no docs yet) Create message report with corrected key presses}
\usage{
get_mesrep(
  mesrep_all,
  fixed_kp,
  final_columns = c("RECORDING_SESSION_LABEL", "CURRENT_MSG_TIME", "TRIAL_INDEX",
    "AudioTarget", "Trial")
)
}
\arguments{
\item{final_columns}{}
}
\description{
(no docs yet) Create message report with corrected key presses
}
