% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/babar.R
\name{get_metrics_and_inventory}
\alias{get_metrics_and_inventory}
\title{Collecting phonetic and consonant inventories as space separated strings, and
canonical related metrics, including:
\itemize{
\item Canonical utterances (per utterance and per syllable): n_canonical/all
\item Canonical babbling ratio: n_canonical/n_canonical + n_non_canonical
\item Total number of syllables
\item Total number of canonical syllables
}}
\usage{
get_metrics_and_inventory(df, minimum_count = 50)
}
\arguments{
\item{df}{babar dataframe}

\item{minimum_count}{the minimum number of occurences of a phoneme per recording
to be counted as a phoneme. Default is 50.}
}
\value{
A dataframe with these new metrics as new columns, one row per recording
}
\description{
Collecting phonetic and consonant inventories as space separated strings, and
canonical related metrics, including:
\itemize{
\item Canonical utterances (per utterance and per syllable): n_canonical/all
\item Canonical babbling ratio: n_canonical/n_canonical + n_non_canonical
\item Total number of syllables
\item Total number of canonical syllables
}
}
