% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_motor_spreadsheet}
\alias{get_motor_spreadsheet}
\title{The the motor questionaire spreadsheet for the SEEDLingS babies}
\usage{
get_motor_spreadsheet(version = NULL, type = "feather")
}
\arguments{
\item{version}{version tag to checkout}

\item{type}{"feather" or "csv". defaults to "feather"}
}
\value{
a tibble contaiing the SEEDLingS Motor Questionaire spreadsheet
}
\description{
The the motor questionaire spreadsheet for the SEEDLingS babies
}
\examples{
motor <- get_motor_spreadsheet(version = '0.0.2')
}
