% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_reliability}
\alias{get_reliability}
\title{Get the reliability spreadsheets}
\usage{
get_reliability(av, month, branch = NULL, commit = NULL)
}
\arguments{
\item{av}{either "audio" or "video"}

\item{month}{the month as a string, e.g. "06"}

\item{branch}{git branch to pull from}

\item{commit}{git commit to pull}
}
\description{
Get the reliability spreadsheets
}
\examples{
audio_06_rel <- get_reliability("audio", "06")
}
