% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_seedlings_nouns}
\alias{get_seedlings_nouns}
\title{Load data from the SEEDLingS - Nouns dataset}
\usage{
get_seedlings_nouns(
  version = NULL,
  table = c("seedlings-nouns", "regions", "recordings", "sub-recordings"),
  get_codebook = FALSE
)
}
\arguments{
\item{version}{version tag to checkout}

\item{table}{Apart from the main "seedlings-nouns" table, the dataset contains three more: regions, recordings, and sub-recordings.
See "public/README.md" of the \href{https://github.com/BergelsonLab/seedlings-nouns_private}{seedlings-nouns_private} for details.}

\item{get_codebook}{Set to \code{TRUE} to get the requested table's codebook instead of the table itself.}
}
\value{
By default, returns a dataframe with one row per annotated object.

If \code{table} and or \code{get_codebook} are changed from their default values, returns the requested table/codebook.
}
\description{
Loads a requested table from the SEEDLingS - Nouns dataset.
By default, loads the main "seedlings-nouns" table with all the annotated nouns in the SEEDLingS corpus.
For the function to work, clone \href{https://github.com/BergelsonLab/seedlings-nouns_private}{seedlings-nouns} to \verb{~/BLAB_DATA/seedlings-nouns/} first.
}
\details{
To get the same data every time you run the script, always supply the version argument.
To get the latest version number, run \code{get_latest_version('seedlings-nouns')} and then set the version parameter to the output number, e.g., \code{get_seedlings_nouns(version = 'v1.0.2')}.

Alternatively, don't set the version parameter, run the function, look for the version number in the issued warning, and then set \code{version} to that number.
You don't need to run the function again after that.

If you are a Bergelson Lab member and you need to use a version that hasn't been made public yet, clone \link{seedlings-nouns_private}(https://github.com/Bergel sonLab/seedlings-nouns_private) to \verb{~/BLAB_DATA/seedlings-nouns_private/}.
If you don't know the version number and want to get the most current one, get it with \code{get_latest_version('seedlings-nouns_private')}.
}
\examples{
seedlings_nouns <- get_seedlings_nouns('0.0.0.9000')

}
