% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_seedlings_nouns}
\alias{get_seedlings_nouns}
\title{Load seedlings_noun from the lab-private repository}
\usage{
get_seedlings_nouns(version = NULL)
}
\arguments{
\item{version}{version tag to checkout}
}
\value{
A dataframe with one row per annotated object.
}
\description{
For the function to work, clone \href{https://github.com/BergelsonLab/seedlings-nouns_private}{seedlings-nouns_private} to \verb{~/BLAB_DATA/seedlings-nouns_private/} if you haven't already.
}
\details{
To make your analysis reproducible, set the \code{version} argument.
To get the latest version, omit \code{version}, look for the version number in the output, and then set the \code{version} argument to that version.
}
\examples{
seedlings_nouns <- get_seedlings_nouns()
}
