% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lena.R
\name{get_seedlings_speaker_stats}
\alias{get_seedlings_speaker_stats}
\title{Calculates the number of Seedlings annotations in each interval (by speaker)}
\usage{
get_seedlings_speaker_stats(intervals, annotations)
}
\arguments{
\item{intervals}{A tibble with required columns \code{interval_start}, \code{interval_end},
and \code{interval_start_wav}. The function expects:
\itemize{
\item \code{interval_start}, \code{interval_end}: POSIXct timestamps with local time
\item \code{interval_start_wav}: Milliseconds since the beginning of the wav file
The function will calculate \code{interval_end_wav} from the timestamp duration.
Can contain other columns.
}}

\item{annotations}{A tibble loaded from a csv annotations file from the
Seedlings project or similar}
}
\value{
A tibble with \code{interval_start} and \code{interval_end} columns for each
interval in \code{intervals} and two new columns: \code{speaker} and \code{n_annotations}.
}
\description{
Calculates the number of Seedlings annotations in each interval (by speaker)
}
