% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lena.R
\name{get_seedlings_speaker_stats}
\alias{get_seedlings_speaker_stats}
\title{Calculates the number of Seedlings annotations in each interval (by speaker)}
\usage{
get_seedlings_speaker_stats(intervals, annotations)
}
\arguments{
\item{intervals}{A tibble with columns depending on \code{time_type}:
\itemize{
\item For \code{wall}: \code{interval_start}, \code{interval_end} (POSIXct), \code{interval_start_wav} (ms)
\item For \code{wav}: \code{interval_start_wav}, \code{interval_end_wav} (both in ms since wav start)
Can contain other columns.
}}

\item{annotations}{A tibble loaded from a csv annotations file from the
Seedlings project or similar}
}
\value{
A tibble with \code{interval_start} and \code{interval_end} columns for each
interval in \code{intervals} and two new columns: \code{speaker} and \code{n_annotations}.
}
\description{
Calculates the number of Seedlings annotations in each interval (by speaker)
}
