% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_vihi_annotations}
\alias{get_vihi_annotations}
\title{Load VIHI annotation data}
\usage{
get_vihi_annotations(
  version = NULL,
  table = c("annotations", "intervals", "merged")
)
}
\arguments{
\item{version}{version tag to checkout}

\item{table}{Which of the two tables should be loaded?}

\item{merge}{Should annotations be merged with the intervals info? Not
applicable if \code{table == 'intervals'}. True by default.}
}
\value{
A tibble with
}
\description{
Clone BLAB-private \href{https://github.com/bergelsonlab/vihi_annotations.git}{vihi_annotations}
repo to \verb{~/BLAB_DATA} once before using this function.
}
\examples{
vihi_annotaitons <- get_vihi_annotations(version='0.0.0.9000', merge=True)
}
