% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lena.R
\name{get_vtc_speaker_stats}
\alias{get_vtc_speaker_stats}
\title{Calculate per-speaker statistics based on the VTC output}
\usage{
get_vtc_speaker_stats(all_rttm, intervals)
}
\arguments{
\item{all_rttm}{An `all.rttm` file from the VTC output loaded with
`read_rttm`.}

\item{intervals}{a tibble with interval_start and interval_end datetime
columns}
}
\value{
A tibble with `interval_start` and `interval_end` columns for each
interval in `intervals` and three new columns: `voice_type` and `duration`
and `count` of VTC annotations.
}
\description{
Calculate per-speaker statistics based on the VTC output
}
