% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eye_tracking.R
\name{get_windows}
\alias{get_windows}
\title{(no docs yet) Add columns placing trials into time windows of short, medium, and long duration}
\usage{
get_windows(
  fix_mes_age,
  bin_size = 20,
  nb_1 = 18,
  short_window_time = DEFAULT_WINDOWS_UPPER_BOUNDS$short,
  med_window_time = DEFAULT_WINDOWS_UPPER_BOUNDS$med,
  long_window_time = DEFAULT_WINDOWS_UPPER_BOUNDS$long
)
}
\arguments{
\item{fix_mes_age}{}

\item{bin_size}{}

\item{nb_1}{}

\item{short_window_time}{}

\item{med_window_time}{}

\item{long_window_time}{}
}
\description{
(no docs yet) Add columns placing trials into time windows of short, medium, and long duration
}
