% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eye_tracking.R
\name{late_target_retrieved}
\alias{late_target_retrieved}
\title{(no docs yet) Load file with manually corrected late target onsets}
\usage{
late_target_retrieved(
  filename,
  drop_list = c("video_pop_time", "video_targetonset", "notes")
)
}
\arguments{
\item{drop_list}{}
}
\description{
(no docs yet) Load file with manually corrected late target onsets
}
