% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lena.R
\name{match_intervals_to_vtc_annotations}
\alias{match_intervals_to_vtc_annotations}
\title{Match intervals to VTC annotations}
\usage{
match_intervals_to_vtc_annotations(intervals, all_rttm)
}
\arguments{
\item{intervals}{A tibble with required columns \code{interval_start}, \code{interval_end},
and \code{interval_start_wav}. The function expects:
\itemize{
\item \code{interval_start}, \code{interval_end}: POSIXct timestamps with local time
\item \code{interval_start_wav}: Milliseconds since the beginning of the wav file
The function will calculate \code{interval_end_wav} from the timestamp duration.
Can contain other columns.
}}

\item{all_rttm}{An \code{all.rttm} file from the VTC output loaded with
\code{read_rttm}.}
}
\description{
Does a brute-force interval join to match intervals to all annotations that
overlap with them.
}
\keyword{internal}
