% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lena.R
\name{match_intervals_to_vtc_annotations}
\alias{match_intervals_to_vtc_annotations}
\title{Match intervals to VTC annotations}
\usage{
match_intervals_to_vtc_annotations(intervals, all_rttm)
}
\arguments{
\item{intervals}{A tibble with columns \code{interval_start}, \code{interval_end},
\code{interval_start_wav}, and \code{recording_id} as output by \code{prepare_intervals}.
Can contain other columns.}

\item{all_rttm}{An \code{all.rttm} file from the VTC output loaded with
\code{read_rttm}.}
}
\description{
Does a brute-force interval join to match intervals to all annotations that
overlap with them.
}
\keyword{internal}
