% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blab_participants.R
\name{parse_raw_id}
\alias{parse_raw_id}
\title{Parse a raw local subject ID and return a cleaned, standardized version of the ID
Will look for a continuous string of letters and underscore (the study name)
followed by a continuous string of digits (the subject number),
and will return a cleaned version of the ID in the format "studyname_subjectnumber",
with the study name in uppercase, all underscore removed, and the subject number as an integer (no leading zeros).
If the input does not match this pattern, it will return the original input.}
\usage{
parse_raw_id(raw_id)
}
\arguments{
\item{raw_id}{A string (or vector of strings when used with \code{dplyr}) representing a raw local subject ID,
which may contain study name and subject number in various formats}
}
\description{
Parse a raw local subject ID and return a cleaned, standardized version of the ID
Will look for a continuous string of letters and underscore (the study name)
followed by a continuous string of digits (the subject number),
and will return a cleaned version of the ID in the format "studyname_subjectnumber",
with the study name in uppercase, all underscore removed, and the subject number as an integer (no leading zeros).
If the input does not match this pattern, it will return the original input.
}
