% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/babar.R
\name{pivot_to_phoneme}
\alias{pivot_to_phoneme}
\title{Wrangling the babar dataframe into long form, one row per phoneme
with sonority value. Whether a sound is considered a phoneme is determined by
the minimum count parameter.}
\usage{
pivot_to_phoneme(df, minimum_count = 50)
}
\arguments{
\item{df}{babar dataframe}

\item{minimum_count}{the minimum number of occurences of a phoneme per recording
to be counted as a phoneme. Default is 50.}
}
\value{
A dataframe with one row per phoneme per recording
}
\description{
Wrangling the babar dataframe into long form, one row per phoneme
with sonority value. Whether a sound is considered a phoneme is determined by
the minimum count parameter.
}
