% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lena.R
\name{prepare_intervals_for_sampling}
\alias{prepare_intervals_for_sampling}
\title{Prepare intervals for sampling}
\usage{
prepare_intervals_for_sampling(intervals, size, duration, allow_fewer)
}
\arguments{
\item{intervals}{A tibble with columns depending on \code{time_type}:
\itemize{
\item For \code{wall}: \code{interval_start}, \code{interval_end} (POSIXct), \code{interval_start_wav} (ms)
\item For \code{wav}: \code{interval_start_wav}, \code{interval_end_wav} (both in ms since wav start)
Can contain other columns.
}}

\item{size}{required sample size}

\item{duration}{Interval duration supported by \code{lubridate::period}, e.g.,
'2 mins'.}

\item{allow_fewer}{Is it OK if there are fewer than size intervals?}
}
\value{
intervals with incomplete intervals removed
}
\description{
\itemize{
\item remove incomplete intervals,
\item check that there are still some left after that,
\item check that there are at least \code{size} intervals, unless \code{allow_fewer} is
\code{TRUE}
}
}
\keyword{internal}
