% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lena.R
\name{prepare_intervals_for_sampling}
\alias{prepare_intervals_for_sampling}
\title{Prepare intervals for sampling}
\usage{
prepare_intervals_for_sampling(intervals, size, duration, allow_fewer)
}
\arguments{
\item{intervals}{A tibble with required columns \code{interval_start}, \code{interval_end},
and \code{interval_start_wav}. The function expects:
\itemize{
\item \code{interval_start}, \code{interval_end}: POSIXct timestamps with local time
\item \code{interval_start_wav}: Milliseconds since the beginning of the wav file
The function will calculate \code{interval_end_wav} from the timestamp duration.
Can contain other columns.
}}

\item{size}{required sample size}

\item{duration}{Interval duration supported by \code{lubridate::period}, e.g.,
'2 mins'.}

\item{allow_fewer}{Is it OK if there are fewer than size intervals?}
}
\value{
intervals with incomplete intervals removed
}
\description{
\itemize{
\item remove incomplete intervals,
\item check that there are still some left after that,
\item check that there are at least \code{size} intervals, unless \code{allow_fewer} is
\code{TRUE}
}
}
\keyword{internal}
