% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lena.R
\name{sample_intervals_periodically}
\alias{sample_intervals_periodically}
\title{Sample intervals periodically, e.g. every hour}
\usage{
sample_intervals_periodically(intervals, duration, period)
}
\arguments{
\item{intervals}{a tibble with interval_start and interval_end datetime
columns}

\item{duration}{Interval duration supported by `lubridate::period`, e.g.,
'2 mins'.}

\item{period}{Period at the end of which we should sample intervals,
e.g., '1 hour'}
}
\value{
A subsample of intervals.
}
\description{
Sample intervals periodically, e.g. every hour
}
