% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lena.R
\name{sample_intervals_periodically}
\alias{sample_intervals_periodically}
\title{Sample intervals periodically, e.g. every hour}
\usage{
sample_intervals_periodically(intervals, duration, period)
}
\arguments{
\item{intervals}{A tibble with required columns \code{interval_start}, \code{interval_end},
and \code{interval_start_wav}. The function expects:
\itemize{
\item \code{interval_start}, \code{interval_end}: POSIXct timestamps with local time
\item \code{interval_start_wav}: Milliseconds since the beginning of the wav file
The function will calculate \code{interval_end_wav} from the timestamp duration.
Can contain other columns.
}}

\item{duration}{Interval duration supported by \code{lubridate::period}, e.g.,
'2 mins'.}

\item{period}{Period at the end of which we should sample intervals,
e.g., '1 hour'}
}
\value{
A subsample of intervals.
}
\description{
Sample intervals periodically, e.g. every hour
}
