% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lena.R
\name{sample_intervals_periodically}
\alias{sample_intervals_periodically}
\title{Sample intervals periodically, e.g. every hour}
\usage{
sample_intervals_periodically(intervals, duration, period)
}
\arguments{
\item{intervals}{A tibble with columns depending on \code{time_type}:
\itemize{
\item For \code{wall}: \code{interval_start}, \code{interval_end} (POSIXct), \code{interval_start_wav} (ms)
\item For \code{wav}: \code{interval_start_wav}, \code{interval_end_wav} (both in ms since wav start)
Can contain other columns.
}}

\item{duration}{Interval duration supported by \code{lubridate::period}, e.g.,
'2 mins'.}

\item{period}{Period at the end of which we should sample intervals,
e.g., '1 hour'}
}
\value{
A subsample of intervals.
}
\description{
Sample intervals periodically, e.g. every hour
}
