% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lena.R
\name{sample_intervals_randomly}
\alias{sample_intervals_randomly}
\title{Sample intervals randomly}
\usage{
sample_intervals_randomly(intervals, size, period, allow_fewer = FALSE,
  seed = NULL)
}
\arguments{
\item{intervals}{a tibble with interval_start and interval_end datetime
columns}

\item{size}{required sample size}

\item{period}{What period (e.g., '5 mins') is the main interval size? Only
intervals of this duration will be sampled.}

\item{allow_fewer}{Is it OK if there are fewer than size intervals?}

\item{seed}{Seed for the random number generator.}
}
\value{
A subsample of intervals.
}
\description{
Sample intervals randomly
}
