% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lena.R
\name{sample_intervals_randomly}
\alias{sample_intervals_randomly}
\title{Sample intervals randomly}
\usage{
sample_intervals_randomly(
  intervals,
  size,
  duration,
  allow_fewer = FALSE,
  seed = NULL
)
}
\arguments{
\item{intervals}{A tibble with columns depending on \code{time_type}:
\itemize{
\item For \code{wall}: \code{interval_start}, \code{interval_end} (POSIXct), \code{interval_start_wav} (ms)
\item For \code{wav}: \code{interval_start_wav}, \code{interval_end_wav} (both in ms since wav start)
Can contain other columns.
}}

\item{size}{required sample size}

\item{duration}{Interval duration supported by \code{lubridate::period}, e.g.,
'2 mins'.}

\item{allow_fewer}{Is it OK if there are fewer than size intervals?}

\item{seed}{Seed for the random number generator.}
}
\value{
A subsample of intervals.
}
\description{
Sample intervals randomly
}
