% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lena.R
\name{sample_intervals_with_highest}
\alias{sample_intervals_with_highest}
\title{Sample intervals that are highest on a given metric}
\usage{
sample_intervals_with_highest(
  intervals,
  column,
  size,
  duration,
  allow_fewer = FALSE
)
}
\arguments{
\item{intervals}{A tibble with required columns \code{interval_start}, \code{interval_end},
and \code{interval_start_wav}. The function expects:
\itemize{
\item \code{interval_start}, \code{interval_end}: POSIXct timestamps with local time
\item \code{interval_start_wav}: Milliseconds since the beginning of the wav file
The function will calculate \code{interval_end_wav} from the timestamp duration.
Can contain other columns.
}}

\item{column}{Name of the column whose value is to be maximized.}

\item{size}{required sample size}

\item{duration}{Interval duration supported by \code{lubridate::period}, e.g.,
'2 mins'.}

\item{allow_fewer}{Is it OK if there are fewer than size intervals?}
}
\value{
A subsample of intervals.
}
\description{
Sample intervals that are highest on a given metric
}
