% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eye_tracking.R
\name{tag_low_data_trials}
\alias{tag_low_data_trials}
\title{Mark "low-data" trials}
\usage{
tag_low_data_trials(
  fixation_timeseries,
  window_column,
  t_start,
  t_end = NULL,
  t_step = 20,
  min_fraction = 1/3
)
}
\arguments{
\item{fixation_timeseries}{A fixations dataframe that is required to minimally contain these columns:
\itemize{
\item \code{recording_id} and \code{trial_index} to uniquely trials.
\item \code{is_good_timepoint} - only rows with TRUE in this column will be counted as having data.
\item Column specified by \code{window_column}.
}}

\item{window_column}{Name of the column that indicates (using factor label "Y") bins that belong to the window which will be tested for insufficient data.}

\item{t_start}{Lower bound of the window of interest in milliseconds from the target onset.}

\item{t_end}{Upper bound of the window of interest in milliseconds from the target onset. If a non-default upper bound was used during the assignment of bins to windows in the \code{get_windows} call, then that value MUST be supplied here. If NULL (default), the default corresponding to the \code{subsetWin} will be used.}

\item{t_step}{The time step in ms. MUST match the one used in \code{fixations_to_timeseries}.}
}
\value{
The input dataframe with boolean \code{is_trial_low_data} column added.
}
\description{
Trials are considered to be "low-data" if less than one third of the window of interest contains fixations on either area of interest. Bins that won't count toward that one third are:
}
\details{
\itemize{
\item Bins that don't have a fixation due to being in a saccade or in a period with lost track. We are assuming that bins in \code{gazeData} were created by \code{binifyFixation} which skips such bins completely - there won't be rows with them.
\item Bins with fixations outside of the areas of interest, including off-screen fixations. We will identify such fixations by NA values in the \code{propt} column.
}

Or looking at it from the opposite perspective, bins that do count toward the one third are bins in \code{gazeData} that:
\itemize{
\item Are present in \code{gazeData}.
\item Are in the window of interest ("Y" in the column specified by \code{subsetWin}).
\item Have fixations on either are of interest (non-NA values in the \code{propt} column).
}
}
